+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `delete_blank` operation
+
 - `delete_images` operation
 
 ## [0.15.0] - 2026-04-19